FileRenamer finished, also passes all tests
Changed SetUp() and TearDown() of FileRenamerTest to occur at each test rather than only once overall.
Slightly modified rename() to throw an exception if the file it is trying to rename to already exists. On the otherhand, revert() doesn't throw an exception and overwrites the file it is being renamed to because revert() should not fail as it is the backup call, plus it shouldn't encounter any file there anyway because that was what it was originally called.